:recycle: Move fixture and consts to `conftest.py`
